feat: 회원가입 필요 없는 부분 삭제 및 수정
