+++ src/App.css
@@ -47,7 +47,7 @@ input:-webkit-autofill,
 input:-webkit-autofill:hover, 
 input:-webkit-autofill:focus, 
 input:-webkit-autofill:active {
-    -webkit-box-shadow: 0 0 0 1000px #f2f2f2 inset;
+    -webkit-box-shadow: 0 0 0 1000px #fff inset;
 }
 
 a {
